# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Connect | Server",
    "version": "2.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "createUser",
        "summary": "Create user",
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "document": {
                    "type": "string"
                  },
                  "birthDate": {
                    "type": "string"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "companyGroup:read",
                        "companyGroup:update",
                        "companyGroup:createPhoneNumber",
                        "companyGroup:deletePhoneNumber",
                        "companyGroup:updatePhoneNumber",
                        "company:create",
                        "company:read",
                        "company:update",
                        "company:delete",
                        "company:createPhoneNumber",
                        "company:deletePhoneNumber",
                        "company:updatePhoneNumber",
                        "unit:create",
                        "unit:read",
                        "unit:update",
                        "unit:delete",
                        "unit:createPhoneNumber",
                        "unit:deletePhoneNumber",
                        "unit:updatePhoneNumber",
                        "base:create",
                        "base:read",
                        "base:update",
                        "base:delete",
                        "base:createPhoneNumber",
                        "base:deletePhoneNumber",
                        "base:updatePhoneNumber",
                        "user:create",
                        "user:read",
                        "user:update",
                        "user:delete",
                        "user:assignRole",
                        "ambulance:create",
                        "ambulance:read",
                        "ambulance:update",
                        "ambulance:delete",
                        "ambulance:switchBase",
                        "ambulance:switchStatus",
                        "ambulanceDocuments:create",
                        "ambulanceDocuments:update",
                        "ambulanceDocuments:delete",
                        "ambulanceDocuments:bulkUpdate",
                        "sys_admin:accessAll"
                      ]
                    }
                  },
                  "companyGroupId": {
                    "type": "string",
                    "format": "cuid",
                    "pattern": "^[cC][^\\s-]{8,}$"
                  },
                  "companiesIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "cuid",
                      "pattern": "^[cC][^\\s-]{8,}$"
                    }
                  },
                  "unitsIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "cuid",
                      "pattern": "^[cC][^\\s-]{8,}$"
                    }
                  },
                  "basesIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "cuid",
                      "pattern": "^[cC][^\\s-]{8,}$"
                    }
                  }
                },
                "required": [
                  "name",
                  "document",
                  "birthDate",
                  "roles",
                  "companyGroupId"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getUsers",
        "summary": "Get users",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "companyGroupId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "companyId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "unitId",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "default": 10,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "perPage",
            "required": false
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "document": {
                            "type": "string"
                          },
                          "avatarUrl": {
                            "nullable": true,
                            "type": "string"
                          },
                          "birthDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "document",
                          "avatarUrl",
                          "birthDate",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "hasPreviousPage": {
                          "type": "boolean"
                        },
                        "currentPage": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "total",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage",
                        "currentPage"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "operationId": "deleteUser",
        "summary": "Soft delete a user",
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companyGroupId": {
                    "type": "string",
                    "format": "cuid",
                    "pattern": "^[cC][^\\s-]{8,}$"
                  }
                },
                "required": [
                  "companyGroupId"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "deletedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateUser",
        "summary": "Update a user's data",
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "avatarUrl": {
                    "nullable": true,
                    "type": "string",
                    "format": "uri"
                  },
                  "birthDate": {},
                  "companyGroupId": {
                    "type": "string",
                    "format": "cuid",
                    "pattern": "^[cC][^\\s-]{8,}$"
                  },
                  "companiesIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "unitIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "roleIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "companyGroup:read",
                        "companyGroup:update",
                        "companyGroup:createPhoneNumber",
                        "companyGroup:deletePhoneNumber",
                        "companyGroup:updatePhoneNumber",
                        "company:create",
                        "company:read",
                        "company:update",
                        "company:delete",
                        "company:createPhoneNumber",
                        "company:deletePhoneNumber",
                        "company:updatePhoneNumber",
                        "unit:create",
                        "unit:read",
                        "unit:update",
                        "unit:delete",
                        "unit:createPhoneNumber",
                        "unit:deletePhoneNumber",
                        "unit:updatePhoneNumber",
                        "base:create",
                        "base:read",
                        "base:update",
                        "base:delete",
                        "base:createPhoneNumber",
                        "base:deletePhoneNumber",
                        "base:updatePhoneNumber",
                        "user:create",
                        "user:read",
                        "user:update",
                        "user:delete",
                        "user:assignRole",
                        "ambulance:create",
                        "ambulance:read",
                        "ambulance:update",
                        "ambulance:delete",
                        "ambulance:switchBase",
                        "ambulance:switchStatus",
                        "ambulanceDocuments:create",
                        "ambulanceDocuments:update",
                        "ambulanceDocuments:delete",
                        "ambulanceDocuments:bulkUpdate",
                        "sys_admin:accessAll"
                      ]
                    }
                  }
                },
                "required": [
                  "companyGroupId"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}